fixing possible syntax bug with childNodes
I tried to implement this code and was getting an error of "Failed to execute 'appendChild' on 'Node'"
By replacing the childNode property with the children property, I was able to successfully run the code.  Not sure if this is truly a bug fix or a workaround related to my own user error in implementing your code, but I thought I'd submit it if it helps make solution easier for others to implement.